Seperate info into pinfo and dinfo
