+++ datumaro/cli/project_commands/__init__.py
@@ -18,12 +18,13 @@ def get_project_commands():
         ("commit", versioning.commit, "Commit changes in tracked files"),
         ("log", versioning.log, "List history"),
         ("status", versioning.status, "Display current status"),
+        ("pinfo", versioning.info, "Print project info"),
         ("", None, ""),
         ("Dataset operations:", None, ""),
         ("diff", dataset_operations.diff, "Compare datasets"),
         ("export", dataset_operations.export, "Export dataset in some format"),
-        ("info", dataset_operations.info, "Print dataset info"),
         ("stats", dataset_operations.stats, "Compute dataset statistics"),
         ("transform", dataset_operations.transform, "Modify dataset items"),
         ("validate", dataset_operations.validate, "Validate dataset"),
+        ("dinfo", dataset_operations.info, "Print dataset info"),
     ]
